[cleanup] Split up BIP135 regtest into separate threshold and grace period tests
This split was initially performed out of necessity while porting this PR to
Bitcoin Core, but it is better to have separate smaller tests - they are long
enough as it is. This will also make it easier to maintain synchronization
between a BU implementation and a Core implementation.
