+++ qa/pull-tester/rpc-tests.py
@@ -242,7 +242,8 @@ testScriptsExt = [ RpcTest(t) for t in [
     'bip68-sequence',
     'bipdersig-p2p',
     'bipdersig',
-    'bip135-genvbvoting-forks',
+    'bip135-grace',
+    'bip135-threshold',
     'getblocktemplate_longpoll',
     'getblocktemplate_proposals',
     'txn_doublespend',
